requirements.txt
streamlit
requests
Pillow
replicate
streamlit-image-select
